==> Pandas/01-pandas_1.ipynb <==
### Pair Programming Pandas I

1. ¿Qué diferencia hay entre una Serie y un DataFrame?


print('Una serie es un array unidimensional que tiene dos tipos de valores, índice y vector, mientras que un dataframe tiene dos dimensiones y en él pueden guardarse todo tipo de datos.')

# Importamos las librerías necesarias:

import pandas as pd
import numpy as np
import random
import os
from os import getcwd

1. Crea dos listas:
- Una que se llame "valores" que tenga 7 números aleatorios.

def lista_generador(n):
      lista = [0]  * n
      for i in range(n):
          lista[i] = random.randint(0, 100)
      return lista

valores = lista_generador(7)



Otra que se llame "indices" que tenga 7 strings que se llamen como los días de la semana.



indices = ['lunes', 'martes', 'miercoles', 'jueves', 'viernes', 'sábado', 'domingo']


2. A partir de estas dos listas cread una Serie donde la variable "valores" sean los valores de la Serie y la variable indices sean los índices de la Serie.


serie1 = pd.Series(valores, indices)

serie1

3. A partir de las dos listas anteriores cread un diccionario usando un bucle for. Para hacer esto tenemos el método zip que nos permite iterar por dos listas a la vez.

diccionario = {}

for i,z in zip(indices, valores):
    print(i, z) # Dejamos esta línea a modo de comprobación.
    diccionario[i] = z


print('El diccionario nos queda tal que así:', diccionario)

4. Crea una Serie con el diccionario creado en el punto anterior.

serie1 = pd.Series(diccionario)
print(serie1)


- ¿Qué forma tiene esta última Serie?
- ¿Qué tipo de datos tiene?
- ¿Qué tamaño tiene?
- ¿Cuáles son los índices de la Serie?
- ¿Y los values?

# Actualizamos nuestra función:

def propiedades(nuestro_array):
    print(f'La forma de nuestro array es', nuestro_array.shape)
    print(f'El tipo de nuestro array es', nuestro_array.dtype)
    print(f'El tamaño de nuestro array es', nuestro_array.size)
    print(f'Los índices de nuestra serie son', nuestro_array.index)
    print(f'Los values de nuestra serie son', nuestro_array.values)

propiedades(serie1)

- Extraed el valor para el índice "lunes".



print(serie1['lunes'])

- Extraed el valor para el índice "lunes", "martes" y "miércoles".

print(serie1[0:3])

- Extraed el valor para el índice "lunes" y "domingo".

# Preguntamos duda

print('Lunes y domingo:', serie1[["lunes","domingo"]])

5. Cread un array de dos dimensiones con 4 filas y 3 columnas.

array1 = np.random.rand(4, 3)
array1

- Convertid el array en un DataFrame donde las columnas se llamen "España", "Francia" e "Italia".

dataframe1 = pd.DataFrame(array1, columns = ["España", "Francia", "Italia"])
dataframe1

- Descargad el archivo "medallas" y cargad el DataFrame:


miruta = os.getcwd()
miruta

nuestro_df = pd.read_csv("datos/01-medallas.csv", index_col = ["País"])

nuestro_df

- Usando el loc extraed todas las columnas de la fila.

# Elegimos la fila "Tailandia".

nuestro_df.loc["Tailandia"]

- Usando el loc extraed todas las columnas de China e Irán.

nuestro_df.loc[["China","Irán"]]

- Usando el loc extraed solo las columnas de "Oro", "Plata" y "Bronce" de China e Irán.


nuestro_df.loc[["China", "Irán"], ["Oro", "Plata", "Bronce"]]

- Usando el iloc extraed la información de la India. Devolved todas las columnas.

nuestro_df.iloc[5]

- Usando el iloc extraed todas las columnas de los paises que estén en posición impar.

nuestro_df.iloc[0::2]

La primera que nos tendría que salir es Japón y la última Corea del Norte.


Lo hacemos de dos maneras, primera:

nuestro_df.iloc[1::]

Segunda:

nuestro_df.iloc[1:10]

print("En ambos casos, es correcto.")

6. Utilizando el código del ejercicio anterior, seleccionad unicamente las columnas "Oro", "Plata" y "Bronce". Contestad a las siguientes preguntas:

- ¿Qué país consiguió más de 200 medallas de oro?


print('El país que consiguió más de 200 medallas de oro fue', nuestro_df.loc[nuestro_df.Oro>200])

- ¿Qué país consiguió más de 200 medallas de oro y más de 1000 de plata?

print('El país que consiguió más de 200 medallas de oro y más de 1000 de plata fue', nuestro_df.loc[(nuestro_df.Oro>200) & (nuestro_df.Plata>1000)])

- ¿Qué país consiguió más de 200 medallas de oro o más de 700 de bronce?

print('El país que consiguió má s de 200 medallas de oro y más de 1000 de plata fue', nuestro_df.loc[(nuestro_df.Oro>200) | (nuestro_df.Bronce>700)])

- Como vemos hay una columna que se llama Unnamed: 0 que no nos interesa. Eliminad esa columna.

nuestro_df_sin_unnamed = nuestro_df.drop('Unnamed: 0', axis = 1)
nuestro_df_sin_unnamed

- Cread una nueva columma que se llame "total" que sea el resultado de la suma de todas las medallas obtenidas por cada país.

nuestro_df_sin_unnamed['Total'] = (nuestro_df["Oro"] + nuestro_df["Plata"] + nuestro_df["Plata"])

nuestro_df_sin_unnamed
