==> Pandas/06-pandas_6.ipynb <==
import pandas as pd
import numpy as np
import os
from os import getcwd
pd.options.display.max_columns = None

###  Pair Programming Pandas VI

# Es el momento de poner en práctica los groupby de pandas. 
# Para esto vamos a trabajar con el csv creado en el ejercicio de pair programming anterior.
# El objetivo de este ejercicio es agrupar los datos para contestar las siguientes preguntas:

df_union = pd.read_csv("datos/05-attacks_union.csv", index_col=0 )# ponemos esto para que no nos agrege mas unnamed
df_union.head(1)

df_union.columns

df_union.shape

df_union.describe().T

### 1. ¿Cuántos ataques hubo por país? Devuelve un DataFrame.
📌 NOTA Ordena los resultados de mayor a menor. 
Esto todavía no lo hemos visto, pero un DataFrame lo podemos ordenar con .sort_values().

df_groupby_pais = pd.DataFrame(df_union.groupby("country")['country'].count())# importante la funcion de agregación
df_groupby_pais.columns = ['ataques']# renombrar la columna porque se llaman igual
df_groupby_pais.sort_values(by='ataques', ascending=False, inplace=True)
df_groupby_pais

### 2. ¿Cuántos ataques ha habido por año? Devuelve un DataFrame.
De nuevo ordena los resultados de mayor a menor.

df_union.head(2)

df_groupby_anio = pd.DataFrame(df_union.groupby("year")['year'].count())
df_groupby_anio.columns = ['ataques']# renombrar la columna porque se llaman igual
df_groupby_anio.sort_values(by='ataques', ascending=False, inplace=True)
df_groupby_anio

### 3. ¿Cuántos ataques hubo por año y sexo? Devolved un DataFrame solo del año 2000 en adelante.

df_union.head(2)

df_groupby_anio = pd.DataFrame(df_union.groupby(["year","sex"])['year'].count())
df_groupby_anio.columns = ['ataques']# renombrar la columna porque se llaman igual
df_groupby_anio.sort_values(by='ataques', ascending=False, inplace=True)
df_groupby_anio

df_groupby_sex = pd.DataFrame(df_union.groupby(["year","sex"])['case_number'].count()).reset_index()
df_groupby_sexo = df_groupby_sex[df_groupby_sex['year']>2000]
df_groupby_sexo.columns = ['year','sex','ataques']
df_groupby_sexo.head(5)

### 4. Primero extraed solo las filas que correspondan a USA, AUSTRALIA y SOUTH AFRICA.
¿Cuántos de los ataques fueron letales por sexo? Devuelve un DataFrame.

df_union.head(2)

df_groupby_paises = df_union[df_union['country'].isin(['USA','AUSTRALIA','SOUTH AFRICA'])]
df_groupby_paises.head(2)

df_fatales = df_groupby_paises[df_groupby_paises["fatal_(y/n)"]=="Y"]
df_fatales.head(2)

df_fatales_sex = df_fatales.groupby('sex')['case_number'].agg(ataques=('count'))
df_fatales_sex = df_fatales_sex.reset_index()
df_fatales_sex

# 5. ¿En qué país hubo más ataques?
Usando el DataFrame creado en el ejercicio anterior. 


df_groupby_paises = df_union[df_union['country'].isin(['USA','AUSTRALIA','SOUTH AFRICA'])].max()
df_groupby_paises.head(2)

# 6. ¿Qué y cuántos tipos de ataques hubo por país?

df_groupby_paises2 = df_union.groupby(["type", "country"])["case_number"].agg(ataques=('count'))
df_ataque_paises = df_groupby_paises2.reset_index()
df_ataque_paises
